% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R,
%   R/pedigree_consistency_checks_tbl_df.R
\name{all_birthdate_consistent}
\alias{all_birthdate_consistent}
\alias{all_birthdate_consistent}
\title{check whether all non-NA birthdates are consistent}
\usage{
all_birthdate_consistent(p_tbl_ped, pn_bd_col_idx, pb_out = FALSE)

all_birthdate_consistent(p_tbl_ped, pn_bd_col_idx, pb_out)
}
\arguments{
\item{p_tbl_ped}{original pedigree as tbl_df}

\item{pn_bd_col_idx}{column index of birthdate}

\item{pb_out}{flag indicating whether debugging output should be written}

\item{p_tbl_ped}{original pedigree as tbl_df}

\item{pn_bd_col_idx}{column index of birthdate}
}
\value{
list of check result and row indices of non-consistent records

list of check result and row indices of non-consistent records
}
\description{
Birthdates are checked using check_birthdate_tbl() and the
resulting validated tbl_df pedigree is compared to the
original pedigree. In case that we find differences, we
return a list of row-indices that are not consistent

Birthdates are checked using check_birthdate_tbl() and the
resulting validated tbl_df pedigree is compared to the
original pedigree. In case that we find differences, we
return a list of row-indices that are not consistent
}
