% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{all_parent_id_consistent}
\alias{all_parent_id_consistent}
\title{Check whether all ids of a given parent (mother or father) have consistent IDs}
\usage{
all_parent_id_consistent(p_tbl_ped, plIdCols, pn_parent_col, p_b_out = FALSE)
}
\arguments{
\item{p_tbl_ped}{original pedigree as tbl_df}

\item{plIdCols}{list with column indices for all ids, required for checking}

\item{pn_parent_col}{column index of parent to be checked}

\item{p_b_out}{flag indicating whether output should be written}
}
\value{
list indicating result of consistency check and list of row indices with inconsistent IDs
}
\description{
Given a pedigree as tbl_df, it is first run through the TVD-ID check using
the function PedigreeFromTvdData::check_tvd_id_tbl(). As a result, we get the
checked pedigree. The original pedigree in p_tbl_ped and the checked pedigree
are compared and the result is returned as a list indicating whether the
two pedigrees are consistent and if not with a vector of row-indices where
inconsistencies did occur.
}
