% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_main_build.R
\name{build_pedigree_from_tvd}
\alias{build_pedigree_from_tvd}
\title{Read K11-formatted TVD-data to construct a pedigree}
\usage{
build_pedigree_from_tvd(ps_tvd_file, ps_out_file = paste(ps_tvd_file, "out",
  sep = "."), pvec_format)
}
\description{
Read K11-formatted TVD-data to construct a pedigree
}
