% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{check_birthdate}
\alias{check_birthdate}
\title{Validation of birthdate border and limit}
\usage{
check_birthdate(plPedigree, lFormatBorder = getBirthdateBorder(),
  lLimitValue = getBirthdayConsistencyLimit())
}
\arguments{
\item{lFormatBorder}{list of consistency border by default taken from getBirthdateBorder()}

\item{lLimitValue}{list of consistency limit value by default taken from getBirthdayConsistencyLimit()}
}
\value{
lCheckedPedigree3
}
\description{
Format of the birdate (YearMonthDay) have to be checked with some limits
}
