% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{check_birthdate_tbl}
\alias{check_birthdate_tbl}
\title{Validation of birthdate format using tbl_df pedigree}
\usage{
check_birthdate_tbl(ptblPedigree, lFormatBorder = getBirthdateBorder(),
  lLimitValue = getBirthdayConsistencyLimit(),
  pnBirthdateColIdx = getBirthdateColIdx())
}
\description{
Validation of birthdate format using tbl_df pedigree
}
