% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{check_parent_as_animal}
\alias{check_parent_as_animal}
\title{Check whether parents appear as animals}
\usage{
check_parent_as_animal(plPedigree)
}
\arguments{
\item{plResultPedigree}{input pedigree}
}
\value{
lCheckResultPedigree with mother and father changed according to checks
}
\description{
Given a pedigree imported from TVD-data, we want to
check whether a given parent also exists as an animal
}
\details{
This section contains details about the function.
}
