% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{check_parent_older_tbl}
\alias{check_parent_older_tbl}
\title{Check that parents are older than offspring}
\usage{
check_parent_older_tbl(ptblPedigree, plIdCols = getTvdIdColsDsch(),
  pnBirthdateColIdx = getBirthdateColIdxDsch())
}
\description{
Check that parents are older than offspring
}
