% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks_tbl_df.R
\name{check_sex_tbl}
\alias{check_sex_tbl}
\title{Validation of sex format using tbl_df pedigree}
\usage{
check_sex_tbl(ptblPedigree, ps_geschlecht = getSexColIdxDsch(),
  lsex = getConsistencySex(), pnTvdIdColIdx = getTvdIdColsDsch())
}
\arguments{
\item{ptblPedigree}{pedigree in tbl_df format}

\item{ps_geschlecht}{get column index of sex}

\item{lsex}{list of consistency values by default taken from getConsistencySex()}

\item{pnTvdIdColIdx}{vector of column indices of TVD-Ids}
}
\description{
Validation of sex format using tbl_df pedigree
}
