% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks_tbl_df.R
\name{check_tvd_id_tbl}
\alias{check_tvd_id_tbl}
\title{Checking Format of TVD-Ids in Pedigree ptblPedigree}
\usage{
check_tvd_id_tbl(ptblPedigree, plFormatBorder = getTVDIdBorder(),
  plIdCols = getTvdIdColsDsch())
}
\arguments{
\item{ptblPedigree}{pedigree as tibble}

\item{lFormatBorder}{list with format borders}

\item{lIdCols}{list with column indices where TVD-ids are stored}
}
\value{
corrected pedigree
}
\description{
Checking Format of TVD-Ids in Pedigree ptblPedigree
}
