% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{check_tvdid}
\alias{check_tvdid}
\title{Validation of tvd-number}
\usage{
check_tvdid(plPedigree, lFormatBorder = getTVDIdBorder())
}
\arguments{
\item{plPedigree}{input pedigree}

\item{lFormatBorder}{list of consistency border by default taken from getTVDIdBorder()}
}
\value{
lCheckedPedigree2 with mother and father changed according to checks
}
\description{
Given a pedigree imported from TVD-data, we want to
check if the first two chars == country, after twelve chars numeric
}
