% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_utilities.R
\name{clean_build_install}
\alias{clean_build_install}
\title{Cleanup, build vignettes, documents and install}
\usage{
clean_build_install(pkg = ".")
}
\arguments{
\item{pkg}{package directory}
}
\description{
Simple utility to clean up old vignette output,
to build vignettes for the current package, to
roxygenize all help-texts and to install the
new version of the package given by pkg.
}
