% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{correct_tvd_format}
\alias{correct_tvd_format}
\title{Correct Ids in column pnIdCol which do not have the correct format}
\usage{
correct_tvd_format(ptblPedigreeResult, plFormatBorder, pnIdCol)
}
\arguments{
\item{ptblPedigreeResult}{input pedigree to be checked as tibble}

\item{plFormatBorder}{list with format borders}

\item{pnIdCol}{column to be checked inside of the pedigree}
}
\value{
tblPedigreeResult corrected tibble pedigree
}
\description{
Given a pedigree in a tibble, set all values in pnIdCol to NA which
do not have the correct format.
}
