% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks_tbl_df.R
\name{correct_tvd_format_tbl}
\alias{correct_tvd_format_tbl}
\title{Correct Ids in column pnIdCol which do not have the correct format tbl}
\usage{
correct_tvd_format_tbl(p_tbl_ped, plFormatBorder = getTVDIdBorder(),
  plIdCols = getTvdIdColsDsch(), pnIdCol = pnIdCol)
}
\arguments{
\item{plFormatBorder}{list with format borders}

\item{pnIdCol}{column to be checked inside of the pedigree}

\item{ptblPedigreeResult}{input pedigree to be checked as tibble}
}
\description{
Correct Ids in column pnIdCol which do not have the correct format tbl
}
