% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getBirthdateColIdx}
\alias{getBirthdateColIdx}
\title{Get column index of birthdate in tbl_df pedigree}
\usage{
getBirthdateColIdx()
}
\value{
column index of birthdate in tbl_df pedigree
}
\description{
Because there are no package-global variables, we
return the fix-coded column index of birthdate as
the result of a function.
}
