% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getBirthdayConsistencyLimit}
\alias{getBirthdayConsistencyLimit}
\title{Get consistency limits for birthdates}
\usage{
getBirthdayConsistencyLimit()
}
\description{
Get consistency limits for birthdates
}
