% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getColPositions}
\alias{getColPositions}
\title{Return default column positions for fixed width coming from TVD}
\usage{
getColPositions()
}
\value{
vector with column positions
}
\description{
Return default column positions for fixed width coming from TVD
}
