% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getDschFormatTblDf}
\alias{getDschFormatTblDf}
\title{Get Format from DSCH xlsx-file as tbl_df}
\usage{
getDschFormatTblDf(psFormatDschFile)
}
\arguments{
\item{sFormatDsch}{path to DSCH-xlsx file}
}
\value{
tbl_df containing DSCH format definitions
}
\description{
The data format defined in DSCH (Datenschnittstelle)
Rindvieh.ch is read from an xlsx file and returned
as a tbl_df. The result contains the complete description
of all formats. From that result single format definitions
can be extracted.
}
