% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getFormatBorder}
\alias{getFormatBorder}
\title{Return default border values for fixed input format}
\usage{
getFormatBorder()
}
\value{
list of format borders
}
\description{
Return default border values for fixed input format
}
