% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getFormatDSCHFile}
\alias{getFormatDSCHFile}
\title{Return path to the DSCH-format file}
\usage{
getFormatDSCHFile()
}
\value{
Path to the DSCH-format file
}
\description{
Return path to the DSCH-format file
}
