% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getK11ColPositionFromDsch}
\alias{getK11ColPositionFromDsch}
\title{Get K11 Format from the DSCH-format file}
\usage{
getK11ColPositionFromDsch(psFormatDschFile, pnTagCol = 2,
  psStartTag = "Satzart K11", psEndTag = NA)
}
\arguments{
\item{psFormatDschFile}{path to DSCH-xlsx file}

\item{pnTagCol}{column index containing tags in DSCH-tbl_df (default = 2)}

\item{psStartTag}{Tag indicating starting row of K11 format (default = "Satzart K11")}

\item{psEndTag}{Tag indicating end row of K11 format (default = NA)}
}
\value{
tbl_df containing K11 definition
}
\description{
Given the DSCH-format read from its xlsx file,
the format definition of K11 data is
extracted and returned as tbl_df
}
