% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getK11ColPositionVecFromDsch}
\alias{getK11ColPositionVecFromDsch}
\title{Get K11 Column positions as vector}
\usage{
getK11ColPositionVecFromDsch(psFormatDschFile, pnColPosition = 5)
}
\arguments{
\item{psFormatDschFile}{name of xlsx-file conaining DSCH-formats}

\item{pnColPosition}{column index where positions are stored}
}
\value{
vector with K11 column positions
}
\description{
Positions where columns are separated in a fixed-width-formatted
input file for the K11 format are taken from the xlsx-DSCH format
definition file. The column separator positions are returned as
a vector of integers, because K11 does not contain any floating
point columns.
}
