% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getSexColIdxDsch}
\alias{getSexColIdxDsch}
\title{Get column index for sex with new DSCH-format}
\usage{
getSexColIdxDsch(psFormatDschFile = getFormatDSCHFile(), pnTagCol = 2,
  ps_pattern_geschlecht = "Geschlecht11")
}
\description{
Get column index for sex with new DSCH-format
}
