% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getTVDIdBorder}
\alias{getTVDIdBorder}
\title{Get consistency border for TVDid}
\usage{
getTVDIdBorder()
}
\value{
list of start and end positions for parts of TVD-ID
}
\description{
The results returned by this function are based
on the definition of a TVD-ID which has two positions
for a country code and twelve positions for a number
}
