% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getTvdIdColsDsch}
\alias{getTvdIdColsDsch}
\title{Get column indices of different IDs in pedigree tbl_df}
\usage{
getTvdIdColsDsch(psFormatDschFile = getFormatDSCHFile(), pnTagCol = 2,
  ps_pattern_tier_id = "Kalb Id", ps_pattern_mutter_id = "Mutter Id",
  ps_pattern_vater_id = "Vater Id")
}
\arguments{
\item{psFormatDschFile}{DSCH format file (default = "RindviehCH_Schnittstelle_4_13_D_K11.xlsx")}

\item{pnTagCol}{column in DSCH format where tags occur}

\item{ps_pattern_tier_id}{Tag for TierID (default = "Kalb Id")}

\item{ps_pattern_mutter_id}{Tag for MutterId (default = "Mutter Id")}

\item{ps_pattern_vater_id}{Tag for VaterId (default = "Vater Id")}
}
\value{
list of column indices for TierId, MutterId and VaterId
}
\description{
Given the DSCH-format of K11 read from psFormatDschFile, the
column index in the peidgree tbl_df is searched using some
tag patterns that are specific for the ID-columns
}
