% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{is.notletter}
\alias{is.notletter}
\title{Check the country-code in TVDid}
\usage{
is.notletter(pId)
}
\arguments{
\item{pId}{Id to be tested}
}
\description{
Check if there are something else than letter in the country-part of tvdid
}
