% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{is.notnumber}
\alias{is.notnumber}
\title{Check the number-code in TVDid}
\usage{
is.notnumber(pId)
}
\description{
Check if there are something else than number after the country letters of tvdid
}
