% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_output.R
\name{output}
\alias{output}
\title{Output Pedigree from TVD-Data}
\usage{
output(plPedigree, psOutfile, psSepChar = "\\t")
}
\arguments{
\item{plPedigree}{processed Pedigree}

\item{psOutfile}{output Pedigree}

\item{psSepChar}{parameter by default for delimiter symbol}
}
\description{
Output Pedigree from TVD-Data
}
