% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedigreeFromTvdData.R
\docType{package}
\name{PedigreeFromTvdData}
\alias{PedigreeFromTvdData}
\alias{PedigreeFromTvdData-package}
\title{PedigreeFromTvdData: A package for importing pedigree data from TVD-input-files.}
\description{
PedigreeFromTvdData: A package for importing pedigree data from TVD-input-files.
}
