% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_input.R
\name{read_tvd_input}
\alias{read_tvd_input}
\title{Reading TVD data and construct a pedigree}
\usage{
read_tvd_input(psInputFile, lFormatBorder = getFormatBorder(),
  pbOut = FALSE)
}
\arguments{
\item{psInputFile}{input file with TVD-daten}

\item{lFormatBorder}{list of format borders}

\item{pbOut}{flag to indicate whether debugging out should be written to the console}
}
\value{
lResultPedigree constructed pedigree as list
}
\description{
First attempt to read a pedigree from a fwf-file and storing
the pedigree-information in a nested list of lists.
}
\section{NOTE}{

The state of this function is deprecated and hence it
should not be used. When reading large files this takes
a very long time to executed and can therefore not be
used for real pedigree data.
}

