% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_input.R
\name{readr_fwf_tvd_input}
\alias{readr_fwf_tvd_input}
\title{Read TVD-Pedigree from file using readr::read_fwf()}
\usage{
readr_fwf_tvd_input(psInputFile, pvecColPosition = getColPositions(),
  pbOut = FALSE)
}
\arguments{
\item{psInputFile}{name of the input file}

\item{pvecColPosition}{vector with column positions}

\item{pbOut}{flag indicating whether output should be written}
}
\value{
tbl_result_pedigree resulting pedigree as tibble
}
\description{
Read TVD-Pedigree from file using readr::read_fwf()
}
