% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_input.R
\name{subBlank}
\alias{subBlank}
\title{Substitute blanks and replace empty string by NA}
\usage{
subBlank(psValue)
}
\arguments{
\item{psValue}{value input}
}
\description{
Substitute blanks and replace empty string by NA
}
