% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_transformation_tbl_df.R
\name{transform_check_birthdate_tbl}
\alias{transform_check_birthdate_tbl}
\title{Transformation of incorrect birthdate format using tbl_df pedigree}
\usage{
transform_check_birthdate_tbl(ptbl_pedigree, output_check, pb_out = FALSE)
}
\arguments{
\item{ptblPedigree}{pedigree in tbl_df format}

\item{lLimitValue}{list with fixed limits for year, month and date}

\item{pnBirthdateColIdx}{column index of birthdates in ptblPedigree}
}
\value{
validated and modified tbl_df pedigree
}
\description{
Transformation of incorrect birthdate format using tbl_df pedigree
}
