% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_transformation_tbl_df.R
\name{transform_check_parent_older_offspring}
\alias{transform_check_parent_older_offspring}
\title{Transformation after check if parents are older than their offspring}
\usage{
transform_check_parent_older_offspring(ptbl_pedigree, output_check,
  pb_keep_dup = FALSE, pb_out = FALSE)
}
\arguments{
\item{ptbl_pedigree}{pedigree in tbl_df format}

\item{output_check}{output of function check_unique_animal_id in tbl_df format}

\item{pb_keep_dup}{should one record of those with duplicate ids be kept}

\item{pb_out}{logfile production with TRUE}
}
\value{
tbl_transform_ped of pedigree records not fullfilling requirements
}
\description{
Given a pedigree in tbl_df format, invalidating birthdate of parent and offspring
where parents are older than their offspring.
}
