% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_transformation_tbl_df.R
\name{transform_check_sex_tbl}
\alias{transform_check_sex_tbl}
\title{Transformation after check of sex format using tbl_df pedigree}
\usage{
transform_check_sex_tbl(ptbl_pedigree, output_check, pb_out = FALSE)
}
\arguments{
\item{ptbl_pedigree}{pedigree in tbl_df format}

\item{output_check}{output of function check_sex_tbl in tbl_df format}

\item{pb_out}{flag whether debugging output should be written}
}
\value{
tbl_transform_ped of pedigree records not fullfilling requirements
}
