% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_transformation_tbl_df.R
\name{transform_correct_tvd_format_tbl}
\alias{transform_correct_tvd_format_tbl}
\title{Transformation after check Correct Ids format in column pnIdCol using tbl}
\usage{
transform_correct_tvd_format_tbl(ptbl_pedigree, output_check, pb_out = FALSE)
}
\arguments{
\item{ptbl_pedigree}{input pedigree to be transformed as tibble}

\item{output_check}{output of function correct_tvd_format_tbl in tbl_df format}

\item{pb_out}{flag whether debugging output should be written}
}
\value{
tbl_transform_ped of pedigree records not fullfilling requirements
}
