% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_transformation_tbl_df.R
\name{transform_unique_animal_id}
\alias{transform_unique_animal_id}
\title{Transformation after heck for uniqueness of Animal-IDs in a pedigree}
\usage{
transform_unique_animal_id(output_check, ptbl_pedigree,
  pn_ani_id_col_idx = getTvdIdColsDsch()$TierIdCol, pb_out = FALSE)
}
