% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_unique_ids.R
\name{unique_id}
\alias{unique_id}
\title{Add unique number to each record}
\usage{
unique_id(plPedigree)
}
\arguments{
\item{plPedigree}{original pedigree}
}
\value{
lProcessedPedigree pedigree with unique ids added
}
\description{
Add unique number to each record
}
