% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_main_build.R
\name{build_check_pedigree_from_tvd}
\alias{build_check_pedigree_from_tvd}
\title{Read K11-formatted TVD-data to build a pedigree}
\usage{
build_check_pedigree_from_tvd(ps_tvd_file,
  pvec_format = getK11ColPositionVecFromDsch())
}
\arguments{
\item{ps_tvd_file}{file containing K11-formatted TVD-data}

\item{pvec_format}{vector with column separator positions of pedigree}
}
\value{
tbl_transform_ped of pedigree records not fullfilling requirements
}
\description{
Read K11-formatted TVD-data to build a pedigree
}
