% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{check_birthdate}
\alias{check_birthdate}
\title{Validation of birthdate}
\usage{
check_birthdate(plPedigree, lFormatBorder = getBirthdateBorder())
}
\value{
lCheckedPedigree3
}
\description{
Validation of birthdate
}
