% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{check_sex}
\alias{check_sex}
\title{Validation of sex}
\usage{
check_sex(plPedigree, lsex = getConsistencySex())
}
\arguments{
\item{lsex}{list of consistency values by default taken from getConsistencySex()}
}
\value{
lCheckedPedigree4
}
\description{
Sex of the parent will be checked
}
