% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks.R
\name{check_sex_tbl}
\alias{check_sex_tbl}
\title{Validation of sex format using tbl_df pedigree}
\usage{
check_sex_tbl(ptblPedigree, lsex = getConsistencySex(),
  pnTvdIdColIdx = getTvdIdCols())
}
\arguments{
\item{ptblPedigree}{pedigree in tbl_df format}

\item{lsex}{list of consistency values by default taken from getConsistencySex()}
}
\description{
Validation of sex format using tbl_df pedigree
}
