% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_consistency_checks_tbl_df.R
\name{check_unique_animal_id}
\alias{check_unique_animal_id}
\title{Check for uniqueness of animal-IDs in tbl_df pedigree}
\usage{
check_unique_animal_id(ptbl_pedigree,
  pn_ani_id_col_idx = getTvdIdColsDsch()$TierIdCol, pb_out = FALSE)
}
\arguments{
\item{ptbl_pedigree}{pedigree in tbl_df format}

\item{pn_ani_id_col_idx}{column index for animal-ID in pedigree}
}
\description{
Given a pedigree in tbl_df-format, we have to make sure
that identifiers of animals (animal-IDs) are unique. This
follows from the general definition of a pedigree and
it is also required, because we want to use the column
as a primary key of the animal's pedigree record.
}
