% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_utilities.R
\name{clean_build_install}
\alias{clean_build_install}
\title{Cleanup, build vignettes, documents and install}
\usage{
clean_build_install(pkg = ".")
}
\description{
Cleanup, build vignettes, documents and install
}
