% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_constant.R
\name{getBirthdateColIdxDsch}
\alias{getBirthdateColIdxDsch}
\title{Get column index for birthdate with new DSCH-format}
\usage{
getBirthdateColIdxDsch(psFormatDschFile = getFormatDSCHFile(), pnTagCol = 2,
  ps_pattern_geburt_dat = "Kalbedatum")
}
\description{
Get column index for birthdate with new DSCH-format
}
