% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_input.R
\name{laf_open_fwf_tvd_input}
\alias{laf_open_fwf_tvd_input}
\title{Read TVD-Pedigree from inputfile using LaF::laf_open_fwf()}
\usage{
laf_open_fwf_tvd_input(ps_input_file, pvec_col_position = getColPositions(),
  pb_out = FALSE)
}
\arguments{
\item{ps_input_file}{name of the input file}

\item{pvec_col_position}{vector with column positions}

\item{pbOut}{flag indicating whether output should be written}
}
\value{
tibble converted to laf containing the pedigree info
}
\description{
Read TVD-Pedigree from inputfile using LaF::laf_open_fwf()
}
