% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PedigreeFromTvdData.R
\docType{package}
\name{PedigreeFromTvdData}
\alias{PedigreeFromTvdData}
\alias{PedigreeFromTvdData-package}
\title{PedigreeFromTvdData: A package for importing pedigree data from TVD-input-files.}
\description{
Package PedigreeFromTvdData is used to construct a pedigree from TVD-input files.
TVD-input files are in a fixed-format according to the specification from the
interface definition called "Datenschnittstelle Rindvieh".
}
\section{Pedigree}{

A pedigree is defined as directed acyclic graph containing the relationship
between parents and offspring.
}

\section{Consistency Checks}{

Besides the parent offspring relationships, there are additional consistency
requirements that can be checked.
}

