% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_input.R
\name{read_tvd_input}
\alias{read_tvd_input}
\title{Reading TVD data and construct a pedigree}
\usage{
read_tvd_input(psInputFile, lFormatBorder = getFormatBorder(),
  pbOut = FALSE)
}
\arguments{
\item{psInputFile}{input file with TVD-daten}

\item{lFormatBorder}{list of format borders}
}
\value{
lResultPedigree constructed pedigree as list
}
\description{
Reading TVD data and construct a pedigree
}
