% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_transformation_tbl_df.R
\name{transform_check_parent_older_offspring}
\alias{transform_check_parent_older_offspring}
\title{Transformation after check if parents are older than their offspring}
\usage{
transform_check_parent_older_offspring(ptbl_pedigree, pn_offspring_col,
  pn_birthday_col, pn_parent_col, pn_date_diff_tol = 10^4)
}
\arguments{
\item{ptbl_pedigree}{pedigree in tbl_df format}

\item{pn_offspring_col}{column index for offspring}

\item{pn_birthday_col}{column index for birthdates of offspring}

\item{pn_parent_col}{column index for parents}

\item{pn_date_diff_tol}{minimum difference between birthdates of parents and offspring}
}
\value{
tbl_df of pedigree records not fullfilling requirements
}
\description{
Given a pedigree in tbl_df format, all parents that are also
present as animals are invalidated, if they are older than
their offspring.
}
