% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_transformation_tbl_df.R
\name{transform_unique_animal_id}
\alias{transform_unique_animal_id}
\title{Transformation after check for uniqueness of Animal-IDs in a pedigree}
\usage{
transform_unique_animal_id(ptbl_pedigree, output_check,
  pn_ani_id_col_idx = getTvdIdColsDsch()$TierIdCol, pb_out = FALSE)
}
\arguments{
\item{ptbl_pedigree}{pedigree in tbl_df format}

\item{output_check}{output of function check_unique_animal_id}

\item{pn_ani_id_col_idx}{column index for animal-ID in pedigree}

\item{pb_out}{flag whether debugging output should be written}
}
\value{
tbl_transform_ped of pedigree records not fullfilling requirements
}
\description{
A given pedigree is transformed such that the resulting pedigree
has only unique animal-IDs. The pedigree-records , we first test whether there are any duplicate records,
using the corresponding check-function for uniqueness of IDs. If
duplicate records are found, we run a series of dplyr::group_by - dplyr::summarise -
dplyr::filter - operations to retain just the unique records.
}
